ci(Update): Refactor database backup directory deletion check.
Simplify the logic for checking and deleting backup directories. This
change removes the unnecessary continue statement, ensuring directories
are deleted only if they exist.

+++ MyExpenses.Wpf/Pages/WelcomePage.xaml.cs
@@ -185,9 +185,7 @@ public partial class WelcomePage
             File.Delete(existingDatabase.FilePath);
 
             var backupDirectory = Path.Join(DbContextBackup.LocalDirectoryBackupDatabase, existingDatabase.FileNameWithoutExtension);
-            if (!Directory.Exists(backupDirectory)) continue;
-
-            Directory.Delete(backupDirectory, true);
+            if (Directory.Exists(backupDirectory)) Directory.Delete(backupDirectory, true);
         }
 
         RefreshExistingDatabases();
